64-bit Windows fixes for nacl_desc.
Added overflow checks and fixed Map and NaClIsNegErrno to deal with
64-bit pointers.

Review URL: http://codereview.chromium.org/399049

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1006 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 80912a86c055a94f3715fbd704913faa90c20a9f

+++ src/trusted/desc/nacl_desc_effector.h
@@ -163,10 +163,10 @@ struct NaClDescEffectorVtbl {
    * at requested sysaddr of nbytes in length.  prot may be
    * NACL_ABI_PROT_NONE for address space squatting purposes.
    */
-  int (*MapAnonymousMemory)(struct NaClDescEffector *vself,
-                            uintptr_t               sysaddr,
-                            size_t                  nbytes,
-                            int                     prot);
+  uintptr_t (*MapAnonymousMemory)(struct NaClDescEffector *vself,
+                                  uintptr_t               sysaddr,
+                                  size_t                  nbytes,
+                                  int                     prot);
 
   /*
    * The bound IMC socket to use as the sender when making a
